main.py
"""
SuperDevAgent - Agente Desarrollador Avanzado de IA

Servidor FastAPI para gestión de agentes IA, selección de modelos,
trazado, evaluación y despliegue.
"""

import os
import uuid
import logging
from typing import Dict, List, Optional, Any
from datetime import datetime

from fastapi import FastAPI, HTTPException, BackgroundTasks
from pydantic import BaseModel
from dotenv import load_dotenv

# Integración con modelos locales
try:
    import ollama
except ImportError:
    ollama = None

# Importar modelo avanzado
try:
    from advanced_model import get_super_model
    super_model = get_super_model()
    ADVANCED_MODEL_AVAILABLE = True
except ImportError:
    super_model = None
    ADVANCED_MODEL_AVAILABLE = False

# Importar agente de monitoreo
try:
    from monitoring_agent import MonitoringAgent
    monitoring = MonitoringAgent()
    MONITORING_AVAILABLE = True
except ImportError:
    monitoring = None
    MONITORING_AVAILABLE = False

# Cargar variables de entorno
load_dotenv()

# Configurar logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Crear aplicación FastAPI
app = FastAPI(
    title="SuperDevAgent API",
    description="API para gestión avanzada de agentes IA",
    version="1.0.0"
)

# Modelos de datos
class AgentCreateRequest(BaseModel):
    name: str
    description: str
    model_type: str = "local"  # local, cloud, azure
    capabilities: List[str] = []

class ModelSelectRequest(BaseModel):
    task: str
    preferences: Dict[str, Any] = {}

class TracingRequest(BaseModel):
    agent_id: str

class EvaluationRequest(BaseModel):
    agent_id: str
    test_cases: List[Dict[str, str]] = []

class DeployRequest(BaseModel):
    agent_id: str
    target: str  # railway, render, vercel, fly

class ParallelsIntegrateRequest(BaseModel):
    vm_name: str
    os: str
    resources: Dict[str, Any] = {}

# Almacenamiento en memoria (para demo)
agents_db: Dict[str, Dict] = {}
models_db: Dict[str, Dict] = {}

# Endpoints

@app.get("/")
async def root():
    """Endpoint raíz"""
    return {"message": "SuperDevAgent API v1.0.0", "status": "running"}

@app.get("/status")
async def status():
    """Estado del servicio"""
    return {
        "status": "running",
        "timestamp": datetime.utcnow().isoformat(),
        "advanced_model": ADVANCED_MODEL_AVAILABLE,
        "monitoring": MONITORING_AVAILABLE,
        "ollama": ollama is not None
    }

@app.post("/agents")
async def create_agent(request: AgentCreateRequest):
    """Crear un nuevo agente IA"""
    agent_id = str(uuid.uuid4())

    agent = {
        "id": agent_id,
        "name": request.name,
        "description": request.description,
        "model_type": request.model_type,
        "capabilities": request.capabilities,
        "created_at": datetime.utcnow().isoformat(),
        "status": "created",
        "tracing_enabled": False,
        "deployed": False
    }

    agents_db[agent_id] = agent

    logger.info(f"Agente creado: {agent_id} - {request.name}")

    return {"agent_id": agent_id, "agent": agent}

@app.post("/models/select")
async def select_model(request: ModelSelectRequest):
    """Seleccionar modelo adecuado para la tarea"""
    model_id = str(uuid.uuid4())

    # Lógica básica de selección de modelo
    if "local" in request.preferences and request.preferences["local"]:
        if ollama:
            selected_model = "llama3.1"
            provider = "ollama"
        else:
            selected_model = "microsoft/DialoGPT-medium"
            provider = "huggingface"
    else:
        selected_model = os.getenv("FOUNDRY_MODEL_DEPLOYMENT_NAME", "gpt-4")
        provider = "azure"

    model = {
        "id": model_id,
        "task": request.task,
        "selected_model": selected_model,
        "provider": provider,
        "preferences": request.preferences,
        "created_at": datetime.utcnow().isoformat()
    }

    models_db[model_id] = model

    logger.info(f"Modelo seleccionado: {selected_model} para tarea: {request.task}")

    return {"model_id": model_id, "model": model}

@app.post("/agents/{agent_id}/tracing")
async def add_tracing(agent_id: str, request: TracingRequest, background_tasks: BackgroundTasks):
    """Añadir trazado a un agente"""
    if agent_id not in agents_db:
        raise HTTPException(status_code=404, detail="Agente no encontrado")

    agent = agents_db[agent_id]
    agent["tracing_enabled"] = True

    # Simular configuración de trazado
    background_tasks.add_task(setup_tracing, agent_id)

    logger.info(f"Trazado habilitado para agente: {agent_id}")

    return {"message": "Trazado configurado", "agent_id": agent_id}

@app.post("/agents/evaluate")
async def evaluate_agent(request: EvaluationRequest, background_tasks: BackgroundTasks):
    """Evaluar rendimiento de un agente"""
    if request.agent_id not in agents_db:
        raise HTTPException(status_code=404, detail="Agente no encontrado")

    agent = agents_db[request.agent_id]

    # Simular evaluación
    evaluation_results = {
        "agent_id": request.agent_id,
        "test_cases_run": len(request.test_cases),
        "passed": len(request.test_cases),  # Simular que pasan todos
        "accuracy": 0.95,
        "timestamp": datetime.utcnow().isoformat()
    }

    background_tasks.add_task(run_evaluation, request.agent_id, request.test_cases)

    logger.info(f"Evaluación iniciada para agente: {request.agent_id}")

    return {"evaluation_id": str(uuid.uuid4()), "results": evaluation_results}

@app.post("/agents/deploy")
async def deploy_agent(request: DeployRequest, background_tasks: BackgroundTasks):
    """Desplegar agente en plataforma especificada"""
    if request.agent_id not in agents_db:
        raise HTTPException(status_code=404, detail="Agente no encontrado")

    agent = agents_db[request.agent_id]

    # Simular despliegue
    deployment_info = {
        "agent_id": request.agent_id,
        "target": request.target,
        "status": "deploying",
        "url": f"https://{request.target}.com/{request.agent_id}",
        "timestamp": datetime.utcnow().isoformat()
    }

    background_tasks.add_task(deploy_to_platform, request.agent_id, request.target)

    agent["deployed"] = True
    agent["deployment_info"] = deployment_info

    logger.info(f"Despliegue iniciado para agente: {request.agent_id} en {request.target}")

    return {"deployment_id": str(uuid.uuid4()), "deployment": deployment_info}

@app.post("/parallels/integrate")
async def integrate_parallels(request: ParallelsIntegrateRequest):
    """Integrar con Parallels Desktop"""
    integration_info = {
        "vm_name": request.vm_name,
        "os": request.os,
        "resources": request.resources,
        "status": "integrated",
        "timestamp": datetime.utcnow().isoformat()
    }

    logger.info(f"Integración Parallels configurada: {request.vm_name}")

    return {"integration_id": str(uuid.uuid4()), "integration": integration_info}

@app.get("/agents")
async def list_agents():
    """Listar todos los agentes"""
    return {"agents": list(agents_db.values())}

@app.get("/agents/{agent_id}")
async def get_agent(agent_id: str):
    """Obtener detalles de un agente"""
    if agent_id not in agents_db:
        raise HTTPException(status_code=404, detail="Agente no encontrado")

    return {"agent": agents_db[agent_id]}

# Funciones auxiliares para tareas en background
async def setup_tracing(agent_id: str):
    """Configurar trazado para un agente"""
    logger.info(f"Configurando trazado para agente {agent_id}")
    # Aquí iría la lógica real de configuración de trazado
    pass

async def run_evaluation(agent_id: str, test_cases: List[Dict[str, str]]):
    """Ejecutar evaluación de agente"""
    logger.info(f"Ejecutando evaluación para agente {agent_id}")
    # Aquí iría la lógica real de evaluación
    pass

async def deploy_to_platform(agent_id: str, target: str):
    """Desplegar agente a plataforma"""
    logger.info(f"Desplegando agente {agent_id} a {target}")
    # Aquí iría la lógica real de despliegue
    pass

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
