requirements.txt
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
python-dotenv
ollama
requests
fastapi
uvicorn
pydantic
